Move toGLSL into GL.Shader & :fire: the Fragment stuff.
