+++ src/GL/Shader.hs
@@ -1,7 +1,9 @@
-{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE GADTs, RankNTypes #-}
 module GL.Shader where
 
 import Control.Exception
+import Data.List (intercalate)
+import Data.Monoid ((<>))
 import Foreign.C.String
 import Foreign.Marshal.Alloc
 import Foreign.Ptr
@@ -9,6 +11,7 @@ import Foreign.Storable
 import GL.Exception
 import Graphics.GL.Core41
 import Graphics.GL.Types
+import Graphics.Shader
 import Prelude hiding (IO)
 
 newtype GLShader = GLShader { unGLShader :: GLuint }
@@ -35,3 +38,21 @@ withCompiledShaders sources body = go sources []
 
 checkShader :: GLShader -> IO GLShader
 checkShader = fmap GLShader . checkStatus glGetShaderiv glGetShaderInfoLog GL_COMPILE_STATUS . unGLShader
+
+
+toGLSL :: Fragment () -> String
+toGLSL shader
+  = pragma "version" "410"
+  <> "out vec4 fragColour;\n"
+  <> main (go shader)
+  where go :: Fragment a -> String
+        go (SetColour c) = "  fragColour = " <> go c <> ";\n"
+        go (V4 x y z w) = "vec4(" <> intercalate ", " (show <$> [ x, y, z, w ]) <> ")"
+        go (V2 x y) = "vec2(" <> intercalate ", " (show <$> [ x, y ]) <> ")"
+        go (Add a b) = go a <> " + " <> go b
+        go (Mul a b) = go a <> " * " <> go b
+        go (Sub a b) = go a <> " - " <> go b
+        go (Div a b) = go a <> " / " <> go b
+        go _ = ""
+        pragma k v = "#" <> k <> " " <> v <> "\n"
+        main body = "void main(void) {\n" <> body <> "}"
